pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hn-digest"
version = "0.1.0"
description = "HackerNews AI Digest App"
authors = [{name = "Kate Silverstein", email = "ksilverstein@mozilla.com"}]
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "sendgrid>=6.10.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]